drf_api_audio/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2==2.9.6
pydub==0.25.1
ffmpeg==1.4